requirements.txt
channels==3.0.4
channels_redis==4.2.1
Django==5.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
matplotlib==3.9.0
matplotlib-inline==0.1.7